repo_miner/pyproject.toml
[project]
name = "pr-repo-minner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "python-decouple>=3.8",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
]
